+++ src/main/kotlin/com/github/jensim/megamanipulator/listener/OnStartListener.kt
@@ -3,16 +3,12 @@ package com.github.jensim.megamanipulator.listener
 import com.github.jensim.megamanipulator.onboarding.OnboardingId
 import com.github.jensim.megamanipulator.onboarding.OnboardingOperator
 import com.github.jensim.megamanipulator.project.MegaManipulatorUtil.isMM
-import com.github.jensim.megamanipulator.settings.MegaManipulatorSettingsState
 import com.github.jensim.megamanipulator.toolswindow.MegaManipulatorTabContentCreator
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.StartupActivity
-import org.slf4j.LoggerFactory
 
 class OnStartListener : StartupActivity {
-
-    private val logger = LoggerFactory.getLogger(javaClass)
 
     override fun runActivity(project: Project) {
         if (isMM(project)) {
@@ -21,11 +17,11 @@ class OnStartListener : StartupActivity {
         } else {
             val factory: MegaManipulatorTabContentCreator = project.service()
             factory.createHelloContent()
-            val settings: MegaManipulatorSettingsState = project.service()
-            if (!settings.seenGlobalOnboarding) {
-                val onboardingOperator: OnboardingOperator = project.service()
+
+            val onboardingOperator: OnboardingOperator = project.service()
+            if (!onboardingOperator.seenGlobalOnboarding()) {
                 onboardingOperator.display(OnboardingId.MM_PROJECT_INSTRUCTION) {
-                    settings.seenGlobalOnboarding = true
+                    onboardingOperator.setSeenGlobalOnboarding()
                 }
             }
         }